lora/requirements.txt
diffusers>=0.11.0
transformers>=4.25.1
scipy
ftfy
fire
wandb
safetensors
opencv-python
torchvision
mediapipe
accelerate
tensorboard